requirements.txt
Django==1.11.7
django-crispy-forms==1.7.2
django-pure-pagination==0.3.0
Markdown==3.1.1
Pillow==6.0.0
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.1
